my_fastapi_project/pyproject.toml
[tool.poetry]
name = "My fastapi project"
version = "0.1.0"
description = ""
authors = ["Tobi-De <tobidegnon@proton.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
pydantic = {extras = ["email", "dotenv"], version = "^1.9.1"}
fastapi-users = "^10.0.6"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.1"}
fastapi-users-tortoise = "^0.2.0"
aerich = "^0.6.3"
arq = "^0.22"
python-multipart = "^0.0.5"
sentry-sdk = "^1.5.12"
rich = "^12.4.4"
typer = "^0.4.1"
watchgod = "^0.8.2"
Jinja2 = "^3.1.2"
httpx = "^0.23.0"
aioaws = "^0.13"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
factory-boy = "^3.2.1"
pre-commit = "^2.19.0"
mypy = "^0.960"
isort = "^5.10.1"
flake8 = "^4.0.1"
poetryup = "^0.7.2"
ipython = "^8.4.0"
aiosmtpd = "^1.4.2"

[tool.aerich]
tortoise_orm = "app.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.isort]
profile = "black"

[tool.mypy]
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
